EWSNet Initialisation

Prepares your R session for communicating with Python. This function first searches your computer for an appropriate Python environment and activates it, importing the vital Python packages required. If no appropriate Python install or environment is found, after asking permission, miniconda is downloaded and an environment created.

Usage

ewsnet_init(envname, pip_ignore_installed = FALSE, conda_refresh = F)

Arguments

envname

A string naming the desired Python environment to create/activate. If no Python or environment found, the functions prompts to install miniconda and the required python packages.

pip_ignore_installed

Boolean. If FALSE, any packages already installed are loaded and not re-downloaded. However, if TRUE, packages are downloaded irregardless, overwriting any version already present (is analagous to updating if required).

conda_refresh

Boolean. If TRUE, miniconda is uninstalled and then gives the user the opportunity to perform a fresh reinstall.

Value

Does not return an object as is simply preparing your R session.

Examples

if (FALSE) bypass_autoinit()
#this function prevents 'reticulate' automatically
#reloading the python environment from the
#previous session. Only needs to be run once
#on first install of 'EWSmethods'.

if (FALSE) ewsnet_init(envname = "EWSNET_env")
#Common errors at this stage result from 'reticulate's
#behaviour. For example, conflicts between 'ewsnet_init'
#and RETICULATE_PYTHON may occur if run inside a
#RStudio R project. To fix this, navigate to
#Preferences -> Python, untick 'Automatically
#activate project-local Python environments'
#and restart R.

if (FALSE) reticulate::py_config()
#If successful, 'EWSNET_env forced by use_python
#function' will be printed.